fixed several warnings in various sources
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@6203 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ registration/include/pcl/registration/impl/correspondence_rejection_features.hpp
@@ -54,7 +54,7 @@ pcl::registration::CorrespondenceRejectorFeatures::getRemainingCorrespondences (
     {
       // Check if the score in feature space is above the given threshold
       // (assume that the number of feature correspondenecs is the same as the number of point correspondences)
-      if (!it->second->isCorrespondenceValid (i))
+      if (!it->second->isCorrespondenceValid (static_cast<int> (i)))
         break;
 
       remaining_correspondences[number_valid_correspondences] = original_correspondences[i];
